fix: moved permission handling to its own package
