+++ composeApp/src/androidMain/kotlin/com/eren76/mangly/permissions/PermissionHandlingNotifications.kt
@@ -1,0 +1,19 @@
+package com.eren76.mangly.permissions
+
+import android.Manifest
+import android.content.Context
+import android.content.pm.PackageManager
+import android.os.Build
+import androidx.core.content.ContextCompat
+
+object NotificationPermissionHandling {
+
+    fun canPostNotifications(context: Context): Boolean {
+        return Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU ||
+            ContextCompat.checkSelfPermission(
+                context,
+                Manifest.permission.POST_NOTIFICATIONS
+            ) == PackageManager.PERMISSION_GRANTED
+    }
+}
+
